requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
idna==3.3
oauthlib==3.1.1
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tweepy==4.4.0
urllib3==1.26.7
